requirements/runtime.txt
atlassian-jwt==1.6
PyBambooHR==0.7.1
Flask-SQLAlchemy==2.2
Flask-SSLify==0.1.5
https://github.com/halkeye/flask_atlassian_connect/archive/master.zip#egg=flask-AtlassianConnect
raven[flask]

